qcacld-3.0: Fix HIF IRQ on RT
The HIF IRQ needs to mask the IRQ line in the hardirq handler and keep it
masked until the tasklet finishes running, at which point the IRQ can be
enabled again. Don't force thread the HIF IRQ on RT and make it safe for
hif_irq_disable() to be called from hardirq context. This splits the HIF
IRQ into threaded and non-threaded parts, where the non-threaded part
simply masks the IRQ line until the tasklet unmasks it later.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: fadlyas07 <mhmmdfdlyas@gmail.com>